Improve Makefile
Fix so that images are only built when the plantuml source changes. To
make this work with the plantuml -> png pattern rule, the sphinx targets
are extracted from the sphinx help output and replace the catch-all
target. This resolves #4.

Make the Makefile self-documenting. Any additional targets with a
"##"-comment at the end of the target line are added to the output of
`make help`.

Add a "clean" target for completeness sake.

Signed-off-by: Chris Hoeppler <christian.hoeppler@bosch.com>