Update build/setup scripts after moving to HL Labs
Previously we deployed the docs website to GitHub organization pages and
used separate repos for the docs source and generated website. Now,
we're going to deploy GitHub project pages from the gh-pages orphan
branch within this repo. This still allows to maintain version control
history for both source and generated website separately. Use the git
worktree feature to concurrently work with both branches checked out in
different directories.
(Cf. https://gohugo.io/hosting-and-deployment/hosting-on-github/ for a
detailed description of that type of setup.)

Update plantuml in the `setup.sh` script to current version which has
svg support.

Signed-off-by: Chris Hoeppler <christian.hoeppler@de.bosch.com>