Add build/revision information to conf & Makefile (#35)
sphinx_rtd_theme's footer template will add build or commit info if
available. Therefore change the configuration and Makefile so that
  - CircleCI's build info or
  - the git commit sha
can picked up and printed on the documentation's footer if available.
The CircleCI environment variables are available if the build is invoked
by CircleCI. The git commit sha will be passed by make to the sphinx
build if it is invoked from a clean git working tree.

The basic build functionality, e.g. the `make html` target still works,
even if invoked from a git archive, i.e. when there is no git
information accessible.

Signed-off-by: Chris Hoeppler <christian.hoeppler@de.bosch.com>